Remofed old codes which are not used in the new structures
