+++ GNUmakefile
@@ -19,7 +19,7 @@ include $(SKOFL_ROOT)/config.gmk
 CXX=g++
 CXXFLAGS += -DNO_EXTERN_COMMON_POINTERS #-DDEBUG
 # if you want to use lates neutrino oscillation parameter, please comment out next line
-CXXFLAGS += -DORIGINAL_NUOSCPARAMETER
+#CXXFLAGS += -DORIGINAL_NUOSCPARAMETER
 FC=gfortran
 FCFLAGS += -w -fPIC -lstdc++
 
@@ -32,6 +32,8 @@ LDFLAGS = $(LOCAL_LIBS) $(LOCAL_INC)
 #LIBSROOT=-L$(ROOTSYS)/lib/ -lCint -lCore -lRIO -lNet -lHist -lGraf -lGraf3d -lGpad -lTree -lRint -lPostscript -lMatrix -lPhysics -lMathCore -lThread -lGui
 
 LN = ln -sf
+
+
 
 #
 #  Objects
@@ -49,7 +51,7 @@ MAINBINS = $(patsubst %.cc, bin/%, $(MAINSRCS))
 SKSNSIMLIBOBJS = $(filter obj/SKSNSim%, $(OBJS))
 SKSNSIMLIBOBJS += $(filter obj/elapseday%, $(OBJS))
 
-main: bin obj bin/main_snburst bin/main_dsnb bin/main_snburst_prev bin/main_dsnb_prev bin/main_dsnb_new bin/main_snburst_new
+main: bin obj bin/main_snburst bin/main_dsnb
 
 library: lib lib/libSKSNSim.so
 
@@ -80,29 +82,14 @@ obj/%.o: src/%.F
 	@echo "[SKSNSim] Building FORTRAN code: $*..."
 	@$(FC) $(FCFLAGS) -c $< -o $@
 
-bin/main_snburst_prev: obj/main_snburst.o $(OBJS)
+bin/main_dsnb: obj/main_dsnb.o $(OBJS)
 	@echo "[SKSNSim] Building executable:	$@..."
 	@LD_RUN_PATH=$(SKOFL) $(CXX) $(CXXFLAGS) -o $@ $^ $(LDLIBS)
 
-bin/main_dsnb_prev: obj/main_dsnb.o $(OBJS)
+bin/main_snburst: obj/main_snburst.o $(OBJS)
 	@echo "[SKSNSim] Building executable:	$@..."
 	@LD_RUN_PATH=$(SKOFL) $(CXX) $(CXXFLAGS) -o $@ $^ $(LDLIBS)
 
-bin/main_dsnb_new: obj/main_dsnb_new.o $(OBJS)
-	@echo "[SKSNSim] Building executable:	$@..."
-	@LD_RUN_PATH=$(SKOFL) $(CXX) $(CXXFLAGS) -o $@ $^ $(LDLIBS)
-
-bin/main_snburst_new: obj/main_snburst_new.o $(OBJS)
-	@echo "[SKSNSim] Building executable:	$@..."
-	@LD_RUN_PATH=$(SKOFL) $(CXX) $(CXXFLAGS) -o $@ $^ $(LDLIBS)
-
-bin/main_snburst: bin/main_snburst_prev
-	@${LN} main_snburst_prev $@
-# @${LN} main_snburst_new $@
-
-bin/main_dsnb: bin/main_dsnb_prev
-	@${LN} main_dsnb_prev $@
-# @${LN} main_dsnb_new $@
 
 lib/libSKSNSim.so: $(SKSNSIMLIBOBJS)
 	@echo "[SKSNSim] Making shared library: $@..."
@@ -114,4 +101,3 @@ obj bin lib:
 clean: 
 	$(RM) -r *.o *~ lib/* src/*~ include/*~ core obj/* bin/* obj bin lib $(TARGET)
 
--include main_dsnb_new.d
